refactor threaded networking & discovery & publishing app logic into their own modules
