+++ wsdiscovery/__init__.py
@@ -1,4 +1,6 @@
-from .daemon import WSDiscovery
 from .qname import QName
 from .scope import Scope
+
+# for backwards compatibility only
+from .discovery import ThreadedWSDiscovery as WSDiscovery
